Download specific versions of Roblox's packages (#5)
# Problem

It was brought up here
https://github.com/flipbook-labs/flipbook/pull/329#issuecomment-3275078043
that there are actually endpoints that can download historic copies of
Roblox's packages, which means it's possible to pin a specific version
to install.

# Solution

To make it possible to pin versions in other projects, the `install`
command now takes a `--version` flag. This flag takes a version string
in the form `0.690.0.6900721`

There is also a new `list` command to get the last several Roblox
versions to make it easy to grab the version string